requirements.txt
imageio==2.13.3
numpy==1.19.5
opencv-python-headless==4.5.4.60
pandas==1.1.5
Pillow==8.4.0
plotly==5.4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.15.0
timm==0.5.4
torch==1.7.1+cu110
torchvision==0.11.1
tqdm==4.46.0